move display component into new package
