+++ site/src/main/kotlin/react/bootstrap/site/components/docs/content/typography/DisplayHeadings.kt
@@ -3,12 +3,12 @@ package react.bootstrap.site.components.docs.content.typography
 import kotlinx.html.SPAN
 import react.RBuilder
 import react.bootstrap.content.tables.table
-import react.bootstrap.content.typography.Display
-import react.bootstrap.content.typography.display
-import react.bootstrap.content.typography.display1
-import react.bootstrap.content.typography.display2
-import react.bootstrap.content.typography.display3
-import react.bootstrap.content.typography.display4
+import react.bootstrap.content.typography.display.Display
+import react.bootstrap.content.typography.display.display
+import react.bootstrap.content.typography.display.display1
+import react.bootstrap.content.typography.display.display2
+import react.bootstrap.content.typography.display.display3
+import react.bootstrap.content.typography.display.display4
 import react.bootstrap.site.components.docs.fixings.SectionComponent
 import react.bootstrap.site.components.docs.fixings.codeExample
 import react.bootstrap.site.components.docs.fixings.liveExample
